Copy outside of N_Viqkdef the vgd_new call, for the case of N_GRID_TYPE_HYBRID_NOTNORM
The previous commit (for the case of N_GRID_TYPE_STAGGERED) dealt with the subroutine,
m_testOneCombo_staggered.  This commit applied the analogous changes to the subroutine,
m_testOneCombo_2GridTypes.

This affects TEST 18 which, in reality, deals with a single grid type.  TEST 22, which
also uses the subroutine m_testOneCombo_2GridTypes, does not yet work.